Move constants to Governance Connector constants
